archive/requirements.txt
pysqueezebox
paho-mqtt
aiomqtt